finally, after 200 years of collective design i found a fucking CoroutineScope and Dispatchers.Default or IO :D and yeah it doesn't block that much anymore

+++ src/main/kotlin/commands/Play.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Play(val guild: Long, val name: String, val cache: Boolean=true, val requester: Double, val channel: Double)
+data class Play(val guild: Long, val name: String, val cache: Boolean=true, val requester: Long, val channel: Long)
