+++ src/main/kotlin/commands/Remove.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Remove(val guild: Double, val index: Int)
+data class Remove(val guild: Long, val index: Int)
