+++ src/main/kotlin/commands/Revind.kt
@@ -3,4 +3,4 @@ package commands
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Revind(val guild: Double)
+data class Revind(val guild: Long)
